+++ apollo-router/src/query_planner/caching_query_planner.rs
@@ -32,12 +32,12 @@ use crate::plugins::progressive_override::LABELS_TO_OVERRIDE_KEY;
 use crate::plugins::telemetry::utils::Timer;
 use crate::query_planner::fetch::SubgraphSchemas;
 use crate::query_planner::BridgeQueryPlannerPool;
-use crate::router_bridge::PlanOptions;
-use crate::router_bridge::UsageReporting;
+use crate::apollo_studio_interop::UsageReporting;
 use crate::services::layers::persisted_queries::PersistedQueryLayer;
 use crate::services::layers::query_analysis::ParsedDocument;
 use crate::services::layers::query_analysis::QueryAnalysisLayer;
 use crate::services::query_planner;
+use crate::services::query_planner::PlanOptions;
 use crate::services::QueryPlannerContent;
 use crate::services::QueryPlannerRequest;
 use crate::services::QueryPlannerResponse;
@@ -678,7 +678,7 @@ mod tests {
     use super::*;
     use crate::json_ext::Object;
     use crate::query_planner::QueryPlan;
-    use crate::router_bridge::UsageReporting;
+    use crate::apollo_studio_interop::UsageReporting;
     use crate::spec::Query;
     use crate::spec::Schema;
     use crate::Configuration;
